+++ concordium-sdk/src/main/java/com/concordium/sdk/ClientV2.java
@@ -9,6 +9,7 @@ import com.concordium.sdk.requests.BlockHashInput;
 import com.concordium.sdk.requests.getaccountinfo.AccountRequest;
 import com.concordium.sdk.responses.BlockIdentifier;
 import com.concordium.sdk.responses.accountinfo.AccountInfo;
+import com.concordium.sdk.responses.blockinfo.BlockInfo;
 import com.concordium.sdk.responses.blocksummary.updates.queues.AnonymityRevokerInfo;
 import com.concordium.sdk.responses.consensusstatus.ConsensusStatus;
 import com.concordium.sdk.responses.cryptographicparameters.CryptographicParameters;
@@ -171,14 +172,30 @@ public final class ClientV2 {
      * @return the cryptographic parameters at the given block.
      * @throws {@link BlockNotFoundException} if the block was not found.
      */
-
-    public CryptographicParameters getCryptographicParameters(final BlockHashInput blockHash) throws BlockNotFoundException {
+    public CryptographicParameters getCryptographicParameters(final BlockHashInput blockHash)
+            throws BlockNotFoundException {
         try {
             var grpcOutput = this.server()
                     .getCryptographicParameters(to(blockHash));
             return to(grpcOutput);
-        } catch (StatusRuntimeException e){
+        } catch (StatusRuntimeException e) {
             throw BlockNotFoundException.from(blockHash.getBlockHash());
+        }
+    }
+    
+    /**
+     * Retrieves a {@link BlockInfo}
+     *
+     * @param blockHashInput the block {@link BlockHashInput} to query.
+     * @return A {@link BlockInfo} for the block
+     * @throws BlockNotFoundException If the block was not found.
+     */
+    public BlockInfo getBlockInfo(BlockHashInput blockHashInput) throws BlockNotFoundException {
+        try {
+            return to(this.server()
+                    .getBlockInfo(to(blockHashInput)));
+        } catch (StatusRuntimeException e) {
+            throw BlockNotFoundException.from(blockHashInput.getBlockHash());
         }
     }